cctools-llvm: add at 15.0.7
cctools-llvm is a replacement for cctools that replaces as much of
cctools with equivalents from LLVM that it can reasonably do. This was
motivated by wanting to reduce dependencies on cctools, which are
updated infrequently by upstream.

To provide a motivating example, the version of `strip` included in
cctools cannot properly strip the archives in compiler-rt in LLVM 15.
Paths are left to bootstrap tools, resulting in failed requisites checks
in the final stdenv build. Since `strip` needs replaced, other tools
were replaced as well.

Note that this has to be done in cctools (or some equivalent) because
some derivations (noteably LLVM) use the bintools of the stdenv directly
instead of going through the wrapper.

The following tools from LLVM are not used in this derivation:

* LLD - not fully compatible with ld64 yet and potentially too big of a change;
* libtool - not a drop-in replacement yet because it does not support
  linker passthrough, which is needed by xcbuild; and
* install_name_tool - fails when trying to build swift-corefoundation.

If other incompatibilities are found, other tools can be reverted.
However, testing for the stdenv bump has shown the others seem to work
pretty well as replacements.

One final caveat: some tools are not duplicated or linked from cctools.
As a rule of thumb, the actual tools on a Darwin system were used to
determine which ones to link, so there are some differences (e.g., in
names or availability).