unzip: fix configure script when using clang 16
Clang 16 makes implicit int and function declarations an error by
default in any mode except for C89. The configure script has two of
these that cause it to misdetect features when using Clang 16.

* Implicit `int main` in the errno check. This is fixed by adding the
  missing `int`.
* Implicit definitions of `opendir` and `closedir`. These are fixed by
  declaring the function definitions directly in `conftest.c`.

For `opendir` and `closedir`, I opted not to include `dirent.h` because
the test is checking which library to link. Whether `dirent.h` can be
used is checked separately. Using this approach preserves the intent of
the check while fixing the incompatibility.