gettext: add separate dev output
gettext must propagate CoreServices, but including it in the default output makes things messy for the Darwin stdenv. Separating it out allows CoreServices to be propagated without stdenv changes other than to use gettext.out in the final allowed requisites list.

Note that the setup hook has been split because it covers two cases: locating gettext translations and making sure libintl is linked. The former should be in the default output while the latter should be in the dev output.