stdenv.darwin: avoid propagating bootstrap packages
If gettext is included in the overlay, it propagates CoreServices from early in the bootstrap, which is undesirable. Packages shouldn’t be getting bootstrap packages propagated to them, and it breaks `overrideSDK`, which assumes it is overriding packages from the same package set (including cross-packages).

Fortunately, nothing depends on gettext in the final stdenv except for gawk and gnutar, which are included in the final stdenv’s `initialPaths`. They can safely be dropped from the overlay because they will be built as needed from the final package set.